+++ src/Controller/CommentController.php
@@ -19,6 +19,7 @@ class CommentController extends AbstractController
      */
     public function index(): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY', null, 'Unable to access this page!');
         $comments = $this->getDoctrine()
             ->getRepository(Comment::class)
             ->findAll();
@@ -33,6 +34,7 @@ class CommentController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY', null, 'Unable to access this page!');
         $comment = new Comment();
         $form = $this->createForm(CommentType::class, $comment);
         $form->handleRequest($request);
@@ -56,6 +58,7 @@ class CommentController extends AbstractController
      */
     public function show(Comment $comment): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY', null, 'Unable to access this page!');
         return $this->render('comment/show.html.twig', [
             'comment' => $comment,
         ]);
@@ -66,6 +69,7 @@ class CommentController extends AbstractController
      */
     public function edit(Request $request, Comment $comment): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY', null, 'Unable to access this page!');
         $form = $this->createForm(CommentType::class, $comment);
         $form->handleRequest($request);
 
@@ -86,6 +90,7 @@ class CommentController extends AbstractController
      */
     public function delete(Request $request, Comment $comment): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY', null, 'Unable to access this page!');
         if ($this->isCsrfTokenValid('delete'.$comment->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($comment);
